Assignment3/requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coverage==6.3.2
cymem==2.0.6
easyocr==1.4.2
en-core-web-sm==3.2.0
idna==3.3
imageio==2.16.2
importlib-metadata==4.11.3
iniconfig==1.1.1
Jinja2==3.1.1
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.6
networkx==2.6.3
numpy==1.21.5
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.3.5
pathy==0.6.1
Pillow==9.1.0
pluggy==1.0.0
preshed==3.0.6
py==1.11.0
pydantic==1.8.2
pyparsing==3.0.7
pytesseract==0.3.9
pytest==7.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.7.3
six==1.16.0
smart-open==5.2.1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.2
thinc==8.0.15
tifffile==2021.11.2
tomli==2.0.1
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
typer==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.9
wasabi==0.9.1
XlsxWriter==3.0.3
zipp==3.8.0
